requirements.txt
pyyaml>=6.0
google-genai>=1.0
git+https://github.com/SaraHan774/ytt.git
